+++ tests/Integration/SolrCloud/CollectionsCurlTest.php
@@ -16,7 +16,8 @@ class CollectionsCurlTest extends AbstractCollectionsTest
         parent::setUp();
         // The default timeout of solarium of 5s seems to be too aggressive on travis and causes random test failures.
         // Set it to the PHP default of 13s.
-        $this->client->setAdapter(new Curl());
-        $this->client->getEndpoint()->setTimeout(CURLOPT_TIMEOUT);
+        $adapter = new Curl();
+        $adapter->setTimeout(CURLOPT_TIMEOUT);
+        $this->client->setAdapter($adapter);
     }
 }