+++ CHANGELOG.md
@@ -29,7 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Zend2HttpAdapter, use PSR-18 http adapter instead
 - GuzzleAdapter, use PSR-18 http adapter instead
 - Guzzle3Adapter, use PSR-18 http adapter instead
-- AdapterHelper
+- Endpoint::setTimeout and Endpoint::getTimeout, configure the timeout on the http adapter instead
 
 
 ## [5.1.6]